+++ sknetwork/linalg/__init__.py
@@ -2,7 +2,7 @@
 from sknetwork.linalg.basics import safe_sparse_dot
 from sknetwork.linalg.eig_solver import EigSolver, LanczosEig
 from sknetwork.linalg.laplacian import get_laplacian
-from sknetwork.linalg.normalization import diagonal_pseudo_inverse, get_norms, normalize
+from sknetwork.linalg.normalizer import diagonal_pseudo_inverse, get_norms, normalize
 from sknetwork.linalg.operators import Regularizer, Laplacian, Normalizer, CoNeighbor
 from sknetwork.linalg.polynome import Polynome
 from sknetwork.linalg.sparse_lowrank import SparseLR