+++ sknetwork/linkpred/nn.py
@@ -11,7 +11,7 @@ from scipy import sparse
 
 from sknetwork.linkpred.base import BaseLinker
 from sknetwork.embedding.base import BaseEmbedding
-from sknetwork.linalg.normalization import normalize
+from sknetwork.linalg.normalizer import normalize
 from sknetwork.utils.check import check_n_neighbors
 from sknetwork.utils.format import get_adjacency
 
@@ -23,11 +23,11 @@ class NNLinker(BaseLinker):
 
     Parameters
     ----------
-    n_neighbors :
+    n_neighbors : int
         Number of nearest neighbors. If ``None``, all nodes are considered.
-    threshold :
+    threshold : float
         Threshold on cosine similarity. Only links above this threshold are kept.
-    embedding_method :
+    embedding_method : :class:`BaseEmbedding`
         Embedding method used to represent nodes in vector space.
         If ``None`` (default), use identity.
 
@@ -95,9 +95,9 @@ class NNLinker(BaseLinker):
 
         Parameters
         ----------
-        input_matrix :
+        input_matrix : sparse.csr_matrix, np.ndarray
             Adjacency matrix or biadjacency matrix of the graph.
-        index :
+        index : np.ndarray
             Index of source nodes to consider. If ``None``, the links are predicted for all nodes.
 
         Returns